- adjusted unit tests - moved factories one directory up
